Removed search from watchOS
